requirements.txt
PyGithub~=1.51
requests~=2.23.0
invalidfire-utils~=2020.10.9.1